removed lfs files

+++ src/test/java/org/rug/simpletests/data/project/VersionTest.java
@@ -13,13 +13,13 @@ public class VersionTest {
     void testRetrievedElementRemoval(){
         var version = antlr.getVersionWith(3);
         var nClassRetrieved = version.getGraph().traversal().V().has("ClassType", "RetrievedClass").count().tryNext().orElse(0L);
-        assertEquals(0L, nClassRetrieved);
+        assertEquals(0L, (long)nClassRetrieved);
         var nPackageRetrieved = version.getGraph().traversal().V().has("PackageType", "RetrievedPackage").count().tryNext().orElse(0L);
-        assertEquals(0L, nPackageRetrieved);
+        assertEquals(0L, (long)nPackageRetrieved);
 
         version = pure.getVersionWith(2);
         var nRetrieved = version.getGraph().traversal().V().has("Type", TextP.containing("retrieved")).count().tryNext().orElse(0L);
-        assertEquals(0, nRetrieved);
+        assertEquals(0, (long)nRetrieved);
     }
 
 }
